Feat/104 refactoring

+++ server/src/main/java/team/pre004/stackoverflowclone/web/controller/AnswerController.java
@@ -1,8 +1,10 @@
 package team.pre004.stackoverflowclone.web.controller;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import team.pre004.stackoverflowclone.domain.post.entity.Answer;
 import team.pre004.stackoverflowclone.domain.post.entity.Question;
@@ -17,19 +19,22 @@ import team.pre004.stackoverflowclone.dto.post.response.LikesDto;
 import team.pre004.stackoverflowclone.handler.ExceptionMessage;
 import team.pre004.stackoverflowclone.handler.ResponseCode;
 import team.pre004.stackoverflowclone.handler.exception.CustomNotAccessItemsException;
+import team.pre004.stackoverflowclone.handler.exception.CustomNotContentByIdException;
 import team.pre004.stackoverflowclone.handler.exception.CustomNotContentItemException;
 import team.pre004.stackoverflowclone.mapper.AnswerMapper;
 import team.pre004.stackoverflowclone.mapper.CommentMapper;
 import team.pre004.stackoverflowclone.security.PrincipalDetails;
 import team.pre004.stackoverflowclone.service.AnswerCommentService;
 import team.pre004.stackoverflowclone.service.AnswerService;
+import team.pre004.stackoverflowclone.service.AuthService;
 import team.pre004.stackoverflowclone.service.CommonService;
 
 import java.util.Set;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/answers")
+@RequestMapping("/api/answers")
 public class AnswerController {
 
     private final AnswerService answerService;
@@ -38,20 +43,13 @@ public class AnswerController {
     private final CommonService commonService;
     private final AnswerCommentService answerCommentService;
     private final UsersRepository usersRepository;
+    private final AuthService authService;
 
-    Users users = Users.builder()
-            .password("4321")
-            .name("답글ward")
-            .email("ward@ward.net")
-            .bio("답글쓰는사람입니다.")
-            .build();
 
     @PostMapping("/{questionId}/add") // 답글 작성 요청
-    public ResponseEntity<?> addAnswer(@PathVariable Long questionId, @RequestBody AnswerPostDto answerPostDto) {
-
-        PrincipalDetails principalDetails = PrincipalDetails.builder().
-                users(usersRepository.save(users))
-                .build();
+    public ResponseEntity<?> addAnswer(@AuthenticationPrincipal PrincipalDetails principalDetails,
+                                       @PathVariable Long questionId,
+                                       @RequestBody AnswerPostDto answerPostDto) {
 
         Answer answer = answerService.save(
                 answerMapper.answerPostDtoToAnswer(
@@ -73,41 +71,45 @@ public class AnswerController {
 
 
     @PutMapping("/{answerId}/edit") // 답글 수정 요청
-    public ResponseEntity<?> editAnswer(@PathVariable Long answerId, @RequestBody AnswerPostDto answerPostDto) {
+    public ResponseEntity<?> editAnswer(@AuthenticationPrincipal PrincipalDetails principalDetails,
+                                        @PathVariable Long answerId,
+                                        @RequestBody AnswerPostDto answerPostDto) {
 
-        PrincipalDetails principalDetails = PrincipalDetails.builder().
-                users(usersRepository.save(users))
-                .build();
-
+        Users user = principalDetails.getOwner();
         if (answerService.findById(answerId).isEmpty()) //해당 게시물이 없을 때 에러메세지
             throw new CustomNotContentItemException(ExceptionMessage.NOT_CONTENT_ANSWER_ID);
 
-        if (principalDetails.getOwner().getOwnerId() != answerId) //접근 유저가 아닐경우
-            throw new CustomNotAccessItemsException(ExceptionMessage.NOT_ACCESS_EDIT_ANSWER);
+        authService.isAuthenticatedUser(user.getOwnerId(), answerService.findById(answerId).get().getOwner().getOwnerId());
 
         Answer answer = answerService.update(answerId, answerMapper.answerPostDtoToAnswer(
                 principalDetails.getOwner(), answerId, answerPostDto
         ));
 
         AnswerInfoDto answerInfoDto = answerMapper.getAnswerInfo(answer);
-
         AnswerRespDto<?> response = AnswerRespDto.builder()
                 .code(ResponseCode.SUCCESS)
                 .answer(answerInfoDto)
                 .message("답글을 변경하였습니다.")
                 .build();
 
+
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
 
-
     @DeleteMapping("/{answerId}") // 답글 삭제 요청
-    public ResponseEntity<?> deleteAnswer(@PathVariable Long answerId) {
+    public ResponseEntity<?> deleteAnswer(@AuthenticationPrincipal PrincipalDetails principalDetails,
+                                          @PathVariable Long answerId) {
+
+        Users user = principalDetails.getOwner();
+        if (answerService.findById(answerId).isEmpty()) //해당 게시물이 없을 때 에러메세지
+            throw new CustomNotContentItemException(ExceptionMessage.NOT_CONTENT_ANSWER_ID);
 
         Question question = answerService.findById(answerId).orElseThrow(
                 () -> new CustomNotContentItemException(ExceptionMessage.NOT_CONTENT_ANSWER_ID)
         ).getQuestion();
+
+        authService.isAuthenticatedUser(user.getOwnerId(), answerService.findById(answerId).get().getOwner().getOwnerId());
 
         answerService.deleteById(answerId);
         Set<Answer> answers = answerService.findAllByQuestion(question);
@@ -121,13 +123,15 @@ public class AnswerController {
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @PostMapping("/{id}/likes-up") // 답글 좋아요 요청
-    public ResponseEntity<?> upAnswerLike(@PathVariable Long id) {
-        Long userId = 1L;
+    @PostMapping("/{answerId}/likes-up") // 답글 좋아요 요청
+    public ResponseEntity<?> upAnswerLike(@PathVariable Long answerId,
+                                          @AuthenticationPrincipal PrincipalDetails principalDetails) {
+
+        Long userId = principalDetails.getOwner().getOwnerId();
 
         LikesDto likes = LikesDto.builder()
-                .likes(answerService.selectLikeUp(userId, id))
-                .postId(id)
+                .likes(answerService.selectLikeUp(userId, answerId))
+                .postId(answerId)
                 .postType(PostType.ANSWER)
                 .build();
 
@@ -139,14 +143,15 @@ public class AnswerController {
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @PostMapping("/{id}/likes-down") // 답글 싫어요 요청
-    public ResponseEntity<?> downAnswerLike(@PathVariable Long id) {
+    @PostMapping("/{answerId}/likes-down") // 답글 싫어요 요청
+    public ResponseEntity<?> downAnswerLike(@PathVariable Long answerId,
+                                            @AuthenticationPrincipal PrincipalDetails principalDetails) {
 
-        Long userId = 1L;
+        Long userId = principalDetails.getOwner().getOwnerId();
 
         LikesDto likes = LikesDto.builder()
-                .likes(answerService.selectLikeDown(userId, id))
-                .postId(id)
+                .likes(answerService.selectLikeDown(userId, answerId))
+                .postId(answerId)
                 .postType(PostType.ANSWER)
                 .build();
 
@@ -159,14 +164,15 @@ public class AnswerController {
 
     }
 
-    @PostMapping("/{id}/likes-up/undo") // 답글 좋아요 요청취소
-    public ResponseEntity<?> upUndoAnswerLike(@PathVariable Long id) {
+    @PostMapping("/{answerId}/likes-up/undo") // 답글 좋아요 요청취소
+    public ResponseEntity<?> upUndoAnswerLike(@PathVariable Long answerId,
+                                              @AuthenticationPrincipal PrincipalDetails principalDetails) {
 
-        Long userId = 1L;
+        Long userId = principalDetails.getOwner().getOwnerId();
 
         LikesDto likes = LikesDto.builder()
-                .likes(answerService.selectLikeUpUndo(userId, id))
-                .postId(id)
+                .likes(answerService.selectLikeUpUndo(userId, answerId))
+                .postId(answerId)
                 .postType(PostType.ANSWER)
                 .build();
 
@@ -178,14 +184,15 @@ public class AnswerController {
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @PostMapping("/{id}/likes-down/undo") // 답글 싫어요 요청 취소
-    public ResponseEntity<?> downUndoAnswerLike(@PathVariable Long id) {
+    @PostMapping("/{answerId}/likes-down/undo") // 답글 싫어요 요청 취소
+    public ResponseEntity<?> downUndoAnswerLike(@PathVariable Long answerId,
+                                                @AuthenticationPrincipal PrincipalDetails principalDetails) {
 
-        Long userId = 1L;
+        Long userId = principalDetails.getOwner().getOwnerId();
 
         LikesDto likes = LikesDto.builder()
-                .likes(answerService.selectLikeDownUndo(userId, id))
-                .postId(id)
+                .likes(answerService.selectLikeDownUndo(userId, answerId))
+                .postId(answerId)
                 .postType(PostType.ANSWER)
                 .build();
 
@@ -197,64 +204,66 @@ public class AnswerController {
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @PostMapping("/{id}/comments") //답글 댓글 작성 요청
-    public ResponseEntity<?> addAnswerComment(@PathVariable Long id, @RequestBody AnswerCommentDto answerCommentDto) {
-
-        PrincipalDetails principalDetails = PrincipalDetails.builder().
-                users(usersRepository.save(users))
-                .build();
+    @PostMapping("/{answerId}/comments") //답글 댓글 작성 요청
+    public ResponseEntity<?> addAnswerComment(@AuthenticationPrincipal PrincipalDetails principalDetails,
+                                              @PathVariable Long answerId,
+                                              @RequestBody AnswerCommentDto answerCommentDto) {
 
         answerCommentService.save(
-                commentMapper.getAnswerComment(principalDetails.getOwner(), id, answerCommentDto));
+                commentMapper.getAnswerComment(principalDetails.getOwner(), answerId, answerCommentDto));
 
         CommentRespDto<?> response = CommentRespDto.builder()
                 .code(ResponseCode.SUCCESS)
-                .comment(commentMapper.getAnswerCommentInfos(answerCommentService.findAllByAnswer(id)))
+                .comment(commentMapper.getAnswerCommentInfos(answerCommentService.findAllByAnswer(answerId)))
                 .build();
 
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @PutMapping("/{id}/comments/{commentId}") //답글 댓글 수정 요청
+    @PutMapping("/{answerId}/comments/{commentId}") //답글 댓글 수정 요청
     public ResponseEntity updateAnswerComment(
-            @PathVariable Long id, @PathVariable Long commentId,
+            @AuthenticationPrincipal PrincipalDetails principalDetails,
+            @PathVariable Long answerId, @PathVariable Long commentId,
             @RequestBody AnswerCommentDto answerCommentDto) {
 
-        PrincipalDetails principalDetails = PrincipalDetails.builder().
-                users(usersRepository.save(users))
-                .build();
+        Users users = principalDetails.getOwner();
 
-         answerCommentService.update(id, commentId, answerCommentDto);
+        authService.isAuthenticatedUser(users.getOwnerId(), answerCommentService.findById(commentId).getOwner().getOwnerId());
+        answerCommentService.update(answerId, commentId, answerCommentDto);
 
 
         CommentRespDto<?> response = CommentRespDto.builder()
                 .code(ResponseCode.SUCCESS)
-                .comment(commentMapper.getAnswerCommentInfos(answerCommentService.findAllByAnswer(id)))
+                .comment(commentMapper.getAnswerCommentInfos(answerCommentService.findAllByAnswer(answerId)))
                 .build();
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    @DeleteMapping("/{id}/comments/{commentId}") //답글 댓글 삭제 요청
-    public ResponseEntity<?> deleteAnswerComment(@PathVariable Long id, @PathVariable Long commentId) {
+    @DeleteMapping("/{answerId}/comments/{commentId}") //답글 댓글 삭제 요청
+    public ResponseEntity<?> deleteAnswerComment(
+            @AuthenticationPrincipal PrincipalDetails principalDetails,
+            @PathVariable Long answerId, @PathVariable Long commentId) {
 
-        answerCommentService.deleteById(id, commentId);
+
+        Users users = principalDetails.getOwner();
+
+        authService.isAuthenticatedUser(users.getOwnerId(), answerCommentService.findById(commentId).getOwner().getOwnerId());
+
+        answerCommentService.deleteById(answerId, commentId);
 
         CommentRespDto<?> response = CommentRespDto.builder()
                 .code(ResponseCode.SUCCESS)
-                .comment(answerCommentService.findAllByAnswer(id))
+                .comment(answerCommentService.findAllByAnswer(answerId))
                 .build();
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
     @PostMapping("/{id}/selected") // 답글 채택 요청
-    public ResponseEntity<?> selectAnswer(@PathVariable Long id) {
+    public ResponseEntity<?> selectAnswer(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable Long id) {
 
-        PrincipalDetails principalDetails = PrincipalDetails.builder().
-                users(usersRepository.save(users))
-                .build();
 
         boolean isAccepted = answerService.acceptAnswer(principalDetails.getOwner().getOwnerId(), id);
 
@@ -270,11 +279,8 @@ public class AnswerController {
     }
 
     @PostMapping("/{id}/selected/undo") //답글 채택 취소 요청
-    public ResponseEntity<?> undoSelectAnswer(@PathVariable Long id) {
+    public ResponseEntity<?> undoSelectAnswer(@AuthenticationPrincipal PrincipalDetails principalDetails, @PathVariable Long id) {
 
-        PrincipalDetails principalDetails = PrincipalDetails.builder().
-                users(usersRepository.save(users))
-                .build();
 
         boolean isAccepted = answerService.acceptAnswerUndo(principalDetails.getOwner().getOwnerId(), id);
